+++ packages/google_sign_in/google_sign_in_android/android/src/test/java/io/flutter/plugins/googlesignin/GoogleSignInTest.java
@@ -42,6 +42,8 @@ public class GoogleSignInTest {
   @Spy MethodChannel.Result result;
   @Mock GoogleSignInWrapper mockGoogleSignIn;
   @Mock GoogleSignInAccount account;
+  @Mock GoogleSignInPlugin.OptionsBuilderFactory optionsBuilderFactory;
+  @Mock GoogleSignInOptions.Builder optionsBuilder;
   @Mock GoogleSignInClient mockClient;
   private GoogleSignInPlugin plugin;
 
@@ -51,10 +53,82 @@ public class GoogleSignInTest {
     when(mockRegistrar.messenger()).thenReturn(mockMessenger);
     when(mockRegistrar.context()).thenReturn(mockContext);
     when(mockRegistrar.activity()).thenReturn(mockActivity);
+    when(optionsBuilderFactory.get(GoogleSignInPlugin.OptionsBuilderFactory.DEFAULT_SIGN_IN))
+        .thenReturn(optionsBuilder);
+    when(optionsBuilderFactory.get(GoogleSignInPlugin.OptionsBuilderFactory.DEFAULT_GAMES_SIGN_IN))
+        .thenReturn(optionsBuilder);
     when(mockContext.getResources()).thenReturn(mockResources);
     plugin = new GoogleSignInPlugin();
     plugin.initInstance(mockRegistrar.messenger(), mockRegistrar.context(), mockGoogleSignIn);
     plugin.setUpRegistrar(mockRegistrar);
+    plugin.setOptionsBuilderFactory(optionsBuilderFactory);
+    plugin.setClientIdIdentifierOverride(0);
+  }
+
+  @Test
+  public void init_PassesForceCodeForRefreshTokenFalseWithClientIdParameter() {
+    HashMap<String, Object> arguments = new HashMap<>();
+    arguments.put("signInOption", GoogleSignInPlugin.OptionsBuilderFactory.DEFAULT_SIGN_IN);
+    arguments.put("requestedScopes", Collections.singletonList("requestedScope"));
+    arguments.put("hostedDomain", null);
+    arguments.put("clientId", "mockClientId");
+    arguments.put("forceCodeForRefreshToken", false);
+
+    MethodCall methodCall = new MethodCall("init", arguments);
+
+    plugin.onMethodCall(methodCall, result);
+
+    verify(optionsBuilder, times(1)).requestServerAuthCode("mockClientId", false);
+  }
+
+  @Test
+  public void init_PassesForceCodeForRefreshTokenTrueWithClientIdParameter() {
+    HashMap<String, Object> arguments = new HashMap<>();
+    arguments.put("signInOption", GoogleSignInPlugin.OptionsBuilderFactory.DEFAULT_SIGN_IN);
+    arguments.put("requestedScopes", Collections.singletonList("requestedScope"));
+    arguments.put("hostedDomain", null);
+    arguments.put("clientId", "mockClientId");
+    arguments.put("forceCodeForRefreshToken", true);
+
+    MethodCall methodCall = new MethodCall("init", arguments);
+
+    plugin.onMethodCall(methodCall, result);
+
+    verify(optionsBuilder, times(1)).requestServerAuthCode("mockClientId", true);
+  }
+
+  @Test
+  public void init_PassesForceCodeForRefreshTokenFalseWithClientIdFromResources() {
+    plugin.setClientIdIdentifierOverride(1);
+    when(mockContext.getString(1)).thenReturn("mockClientId");
+    HashMap<String, Object> arguments = new HashMap<>();
+    arguments.put("signInOption", GoogleSignInPlugin.OptionsBuilderFactory.DEFAULT_SIGN_IN);
+    arguments.put("requestedScopes", Collections.singletonList("requestedScope"));
+    arguments.put("hostedDomain", null);
+    arguments.put("forceCodeForRefreshToken", false);
+
+    MethodCall methodCall = new MethodCall("init", arguments);
+
+    plugin.onMethodCall(methodCall, result);
+
+    verify(optionsBuilder, times(1)).requestServerAuthCode("mockClientId", false);
+  }
+
+  @Test
+  public void init_PassesForceCodeForRefreshTokenTrueWithClientIdFromResources() {
+    plugin.setClientIdIdentifierOverride(1);
+    when(mockContext.getString(1)).thenReturn("mockClientId");
+    HashMap<String, Object> arguments = new HashMap<>();
+    arguments.put("signInOption", GoogleSignInPlugin.OptionsBuilderFactory.DEFAULT_SIGN_IN);
+    arguments.put("requestedScopes", Collections.singletonList("requestedScope"));
+    arguments.put("hostedDomain", null);
+    arguments.put("forceCodeForRefreshToken", true);
+
+    MethodCall methodCall = new MethodCall("init", arguments);
+
+    plugin.onMethodCall(methodCall, result);
+
+    verify(optionsBuilder, times(1)).requestServerAuthCode("mockClientId", true);
   }
 
   @Test